+++ R/Studio_DocumentationPage.R
@@ -128,15 +128,18 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
             if ("TextEmbeddingModel" %in% class(model)) {
               shiny::removeModal()
               return(model)
-            } else {
-              display_errors(
-                title = "Error",
-                size = "l",
-                easy_close = TRUE,
-                error_messages = "The file does not contain an object of class TextEmbeddingModel."
-              )
-              return(NULL)
-            }
+            } else if ("BaseModelCore" %in% class(model)) {
+                shiny::removeModal()
+                return(model)
+              } else {
+                display_errors(
+                  title = "Error",
+                  size = "l",
+                  easy_close = TRUE,
+                  error_messages = "The file does not contain an object of class of TextEmbeddingModel or BaseModelCore."
+                )
+                return(NULL)
+              }
           } else if (type == "Classifier") {
             if (
               "TEClassifierRegular" %in% class(model) |
@@ -228,8 +231,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
         citation = input[[paste0("doc_", "developed_by", "_citation")]],
         url = input[[paste0("doc_", "developed_by", "_url")]]
       )
-      r_interface_path <- paste0(model_path(), "/r_interface.rda")
-      save(tmp_model, file = r_interface_path)
+      r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+             tmp_config=create_config_state(tmp_model)
+             save(tmp_config, file = r_interface_path)
       model <- shiny::reactive({
         tmp_model
       })
@@ -269,8 +273,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
           citation = input[[paste0("doc_", "modified_by", "_citation")]],
           url = input[[paste0("doc_", "modified_by", "_url")]]
         )
-        r_interface_path <- paste0(model_path(), "/r_interface.rda")
-        save(tmp_model, file = r_interface_path)
+       r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+               tmp_config=create_config_state(tmp_model)
+               save(tmp_config, file = r_interface_path)
         model <- shiny::reactive({
           tmp_model
         })
@@ -293,8 +298,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
         abstract_eng = input$doc_editor_abstract_eng,
         keywords_eng = input$doc_editor_abstract_eng_keywords
       )
-      r_interface_path <- paste0(model_path(), "/r_interface.rda")
-      save(tmp_model, file = r_interface_path)
+      r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+      tmp_config=create_config_state(tmp_model)
+      save(tmp_config, file = r_interface_path)
       model <- shiny::reactive({
         tmp_model
       })
@@ -319,8 +325,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
       tmp_model$set_model_description(
         eng = input$doc_editor_description_eng
       )
-      r_interface_path <- paste0(model_path(), "/r_interface.rda")
-      save(tmp_model, file = r_interface_path)
+     r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+             tmp_config=create_config_state(tmp_model)
+             save(tmp_config, file = r_interface_path)
       model <- shiny::reactive({
         tmp_model
       })
@@ -346,8 +353,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
         abstract_native = input$doc_editor_abstract_native,
         keywords_native = input$doc_editor_abstract_native_keywords
       )
-      r_interface_path <- paste0(model_path(), "/r_interface.rda")
-      save(tmp_model, file = r_interface_path)
+     r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+             tmp_config=create_config_state(tmp_model)
+             save(tmp_config, file = r_interface_path)
       model <- shiny::reactive({
         tmp_model
       })
@@ -373,8 +381,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
       tmp_model$set_model_description(
         native = input$doc_editor_description_native
       )
-      r_interface_path <- paste0(model_path(), "/r_interface.rda")
-      save(tmp_model, file = r_interface_path)
+     r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+             tmp_config=create_config_state(tmp_model)
+             save(tmp_config, file = r_interface_path)
       model <- shiny::reactive({
         tmp_model
       })
@@ -400,8 +409,9 @@ DocumentPage_Server <- function(id, volumes, type = "TextEmbeddingModel") {
       tmp_model$set_documentation_license(input$doc_editor_documentation_license)
       tmp_model$set_model_license(input$doc_editor_software_license)
 
-      r_interface_path <- paste0(model_path(), "/r_interface.rda")
-      save(tmp_model, file = r_interface_path)
+     r_interface_path <- paste0(model_path(), "/r_config_state.rda")
+            tmp_config=create_config_state(tmp_model)
+            save(tmp_config, file = r_interface_path)
       model <- shiny::reactive({
         tmp_model
       })