+++ R/Studio_Tabs_Training.R
@@ -153,11 +153,18 @@ Training_Server <- function(id, model) {
       {
         shiny::req(model)
         if ("TextEmbeddingModel" %in% class(model())) {
+          plot <- model()$BaseModel$plot_training_history(
+            y_min = input$y_min,
+            y_max = input$y_max,
+            text_size=input$text_size
+          )
+        } else if(("BaseModelCore" %in% class(model()))){
           plot <- model()$plot_training_history(
             y_min = input$y_min,
-            y_max = input$y_max
+            y_max = input$y_max,
+            text_size=input$text_size
           )
-        } else if ("ClassifiersBasedOnTextEmbeddings" %in% class(model())) {
+        }  else if ("ClassifiersBasedOnTextEmbeddings" %in% class(model())) {
           # Necessary input
           shiny::req(input$measure)